+++ lib/Controller/AdminController.php
@@ -7,7 +7,6 @@ namespace OCA\NcTower\Controller;
 use OCA\NcTower\AppInfo\Application;
 use OCA\NcTower\Service\EndpointConfigService;
 use OCP\AppFramework\Controller;
-use OCP\AppFramework\Http\Attribute\AdminRequired;
 use OCP\AppFramework\Http\JSONResponse;
 use OCP\IRequest;
 
@@ -19,7 +18,6 @@ class AdminController extends Controller {
 		parent::__construct(Application::APP_ID, $request);
 	}
 
-	#[AdminRequired]
 	public function saveSettings(): JSONResponse {
 		$this->endpoints->saveFromRequest($this->request->getParams());
 		return new JSONResponse([