+++ templates/userlist.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use OCP\Util;
-
-Util::addScript(OCA\NcTower\AppInfo\Application::APP_ID, 'nc_tower-app');
-?>
-<div id="nc_tower" data-page="users"></div>